refactor(base-gen): extract UsageBillingMixin (Tier-1b Task 6)
Move the money-critical Bucket A (usage-event + ② Learning billing helpers)
verbatim into base_generation/_usage_billing.py: _usage_pipeline,
_usage_context, _record_generation_event, _extraction_input_text,
_record_billing_learning_events, and the pure _count_generated_results
staticmethod. Compose UsageBillingMixin ahead of the existing mixins.

Billing emit-ordering + resets stay on the base in _run_generation; the
moved helpers only READ the accumulators via self. MRO. SINK-3: the
billing_meter/billing_signals imports stay function-local so the drain
test's source-module patch seam survives. Both exception swallows and the
double-bill guard are preserved verbatim. Stubs are annotation-only.
